added no tf testing during training
