+++ common/train_arguments.py
@@ -64,6 +64,14 @@ _train_parser.add_argument(
     '--tokenizer', nargs='+', default=None,
     type=str, help='Tokenizer(s) in pretrained to use.'
 )
+_train_parser.add_argument(
+    '--test_freq', default=None,
+    type=int, help='Frequency of no tf testing.'
+)
+_train_parser.add_argument(
+    '--test_batches', default=50,
+    type=int, help='Number of batches to test without tf.'
+)
 
 ##### distributed training args
 _train_parser.add_argument(
